+++ internal/auth/auth.go
@@ -1,4 +1,4 @@
-package handlers
+package auth
 
 import (
 	"context"
