+++ internal/components/layout/types.go
@@ -1,5 +1,10 @@
 package layout
 
+import (
+	"july/internal/auth"
+	"net/http"
+)
+
 type NavItem struct {
 	Label string
 	URL   string
@@ -14,4 +19,15 @@ type LayoutData struct {
 type UserInfo struct {
 	Username  string
 	AvatarURL string
+}
+
+func UserInfoFromContext(r *http.Request) *UserInfo {
+	u := auth.UserFromContext(r.Context())
+	if u == nil {
+		return nil
+	}
+	return &UserInfo{
+		Username:  u.Username,
+		AvatarURL: u.AvatarURL,
+	}
 }
