+++ internal/features/blog/handler.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 
 	"july/internal/components/layout"
-	"july/internal/handlers"
 
 	"github.com/rs/zerolog/log"
 )
@@ -38,7 +37,7 @@ func (h *Handler) List(w http.ResponseWriter, r *http.Request) {
 	layout := layout.LayoutData{
 		Title:       "Blog",
 		CurrentPath: "/blog",
-		User:        userInfoFromContext(r),
+		User:        layout.UserInfoFromContext(r),
 	}
 
 	BlogListPage(layout, posts).Render(ctx, w)
@@ -60,19 +59,8 @@ func (h *Handler) Detail(w http.ResponseWriter, r *http.Request) {
 	layout := layout.LayoutData{
 		Title:       post.Title,
 		CurrentPath: fmt.Sprintf("/blog/%s", slug),
-		User:        userInfoFromContext(r),
+		User:        layout.UserInfoFromContext(r),
 	}
 
 	BlogPostPage(layout, post).Render(ctx, w)
-}
-
-func userInfoFromContext(r *http.Request) *layout.UserInfo {
-	u := handlers.UserFromContext(r.Context())
-	if u == nil {
-		return nil
-	}
-	return &layout.UserInfo{
-		Username:  u.Username,
-		AvatarURL: u.AvatarURL,
-	}
 }
