+++ internal/features/projects/analysis.go
@@ -1,4 +1,4 @@
-package handlers
+package projects
 
 import (
 	"encoding/json"
@@ -9,6 +9,7 @@ import (
 	"github.com/jackc/pgx/v5"
 	"github.com/rs/zerolog/log"
 
+	"july/internal/auth"
 	"july/internal/db"
 	"july/internal/metrics"
 	"july/internal/shared"
@@ -34,7 +35,7 @@ func (h *ProjectHandler) PostProjectAnalysis(w http.ResponseWriter, r *http.Requ
 	ctx := r.Context()
 	projectID := r.PathValue("projectID")
 
-	sessionUser := UserFromContext(ctx)
+	sessionUser := auth.UserFromContext(ctx)
 	if sessionUser == nil {
 		http.Error(w, "unauthorized", http.StatusUnauthorized)
 		return
