+++ internal/features/projects/analysis_test.go
@@ -1,4 +1,4 @@
-package handlers_test
+package projects_test
 
 import (
 	"context"
