+++ internal/features/projects/permissions.go
@@ -1,4 +1,4 @@
-package handlers
+package projects
 
 import (
 	"fmt"
