+++ internal/features/projects/permissions_test.go
@@ -1,4 +1,4 @@
-package handlers
+package projects
 
 import (
 	"testing"
