+++ internal/features/projects/rescan_test.go
@@ -1,4 +1,4 @@
-package handlers_test
+package projects_test
 
 import (
 	"net/http"
