+++ internal/handlers/userinfo.go
@@ -1,6 +1,7 @@
 package handlers
 
 import (
+	"july/internal/auth"
 	"july/internal/components/layout"
 	"net/http"
 )
@@ -8,7 +9,7 @@ import (
 // getUserFromContext wraps handlers.UserFromContext and converts to *layout.UserInfo.
 // This is used by handlers that haven't yet been migrated to the features/ layout.
 func getUserFromContext(r *http.Request) *layout.UserInfo {
-	u := UserFromContext(r.Context())
+	u := auth.UserFromContext(r.Context())
 	if u == nil {
 		return nil
 	}
